ãfeat: 소셜로그인 오류 안남 ㅠㅠ

+++ src/main/java/com/example/demo/src/user/service/UserSignUpAndFindService.java
@@ -1,16 +1,14 @@
 package com.example.demo.src.user.service;
 
+import com.example.demo.common.exceptions.BaseException;
+import com.example.demo.common.response.BaseResponseStatus;
+import com.example.demo.src.refresh.domain.RefreshToken;
 import com.example.demo.src.user.dao.UserMapper;
 import com.example.demo.src.user.domain.User;
-import com.example.demo.src.user.dto.UserResponseDto;
 import com.example.demo.src.user.model.Role;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.util.Optional;
-
-import static com.example.demo.src.user.dto.UserResponseDto.*;
 
 
 /*
@@ -26,21 +24,19 @@ public class UserSignUpAndFindService {
     @Autowired
     private final UserMapper userMapper;
 
-    // 회원 저장 -> 더 고민하기
-    public void signUp(User user) {
-        if (userMapper.isExistByNickname(user.getUserNickname())) {
-            // 이미 회원인 경우 -> 메인페이지로 리다이렉트
-        }
-        // 신규회원인경우 -> 닉네임 설정 페이지로 리다이렉트
+    /*
+    회원가입관련은 CustoOAuth2UserService에 구현
+    * */
+    public void save(User user) {
         userMapper.save(user);
     }
 
+
     public void setUserNickname(User user, String nickname) { // user는 토큰 공부 더 하고 수정예정
         if (isExistByNickname(nickname)) {
-            throw new IllegalStateException("이미 존재하는 닉네임입니다");
+            throw new BaseException(BaseResponseStatus.ALREADY_USE_NICKNAME);
         } else {
             user.setUserNickname(nickname);
-
         }
     }
 
@@ -51,27 +47,21 @@ public class UserSignUpAndFindService {
     // ID로 회원 찾기
     public User findById(long userId) {
         return userMapper.findById(userId)
-                .orElseThrow(() -> new NullPointerException("해당 유저가 존재하지 않습니다."));
-    }
-
-    // provideId로 회원찾기
-    public User findByUserProvideId(String userProvideId) {
-        return userMapper.findByProvideId(userProvideId)
-                .orElseThrow(() -> new NullPointerException("해당 유저가 존재하지 않습니다"));
+                .orElseThrow(() -> new BaseException(BaseResponseStatus.NOT_FOUND_USER));
     }
 
 
     // Nickname으로 회원찾기
     public User findByUserNickname(String userNickname) {
         return userMapper.findByNickname(userNickname)
-                .orElseThrow(() -> new NullPointerException("해당 유저가 존재하지 않습니다."));
+                .orElseThrow(() -> new BaseException(BaseResponseStatus.NOT_FOUND_USER));
     }
 
     
     // Nickname으로 회원 삭제
     public void deleteByNickname(String userNickname) {
-        userMapper.deleteByNickname(userNickname)
-                .orElseThrow(() -> new NullPointerException("해당 유저가 존재하지 않습니다."));
+        if (isExistByNickname(userNickname)) throw new BaseException(BaseResponseStatus.NOT_FOUND_USER);
+        userMapper.deleteByNickname(userNickname);
 
     }
 
@@ -82,6 +72,18 @@ public class UserSignUpAndFindService {
     // 메일 이미 있는지 여부
     public boolean isExistByEmail(String email) {
         return userMapper.isExistByEmail(email);
+    }
+
+
+    public User findByProvideId(String userProvideId) {
+        return userMapper.findByProvideId(userProvideId)
+                .orElseThrow(() -> new BaseException(BaseResponseStatus.NOT_FOUND_USER));
+    }
+
+
+    // 이건 소셜로그인용
+    public User checkByProvideId(String userProvideId) {
+        return userMapper.findByProvideId(userProvideId).orElse(null);
     }
